requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.19.2
opencv-python-headless==4.4.0.44
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
scipy==1.5.2
